+++ src/commands/datapackCommand.ts
@@ -2,16 +2,14 @@ import * as vscode from 'vscode';
 
 import VlocodeService from '../services/vlocodeService';
 import VlocityDatapackService, * as vds from '../services/vlocityDatapackService';
-import * as s from '../singleton';
-import { isError } from 'util';
-import * as path from 'path';
-import { Command } from "./command";
-import { fstatAsync, readdirAsync, unique } from '../util';
+import * as serviceProvider from '../singleton';
+import { CommandBase } from "./commandBase";
+import { unique } from '../util';
 
-export abstract class DatapackCommand extends Command {
+export abstract class DatapackCommand extends CommandBase {
 
     protected get datapackService() : VlocityDatapackService {
-        return s.get(VlocodeService).datapackService;
+        return serviceProvider.get(VlocodeService).datapackService;
     }
 
     protected async resolveDatapackHeaders(files: vscode.Uri[], reportErrors: boolean = true) : Promise<vscode.Uri[]> {
@@ -27,7 +25,7 @@ export abstract class DatapackCommand extends Command {
             let unresolvedFiles = results.filter(v => !v.header);
             if (unresolvedFiles.length > 0) {
                 let errorMessageText = `${unresolvedFiles.length} of the selected files ${unresolvedFiles.length === 1 ? 'is' : 'are'} not part of a Vlocity datapack. \nSee the log for details.`;
-                vscode.window.showWarningMessage(errorMessageText, { title: 'View log...' }).then(o => o && s.get(VlocodeService).focusLog());
+                vscode.window.showWarningMessage(errorMessageText, { title: 'View log...' }).then(o => o && serviceProvider.get(VlocodeService).focusLog());
                 unresolvedFiles.forEach(f => this.logger.warn(`Unabled to resolve datapack header for: ${f.file.fsPath}`));
             }
         }