+++ src/constants.ts
@@ -2,7 +2,7 @@
  * Default constants used in this extensipn
  */
 
-export const VERSION = '0.5.2';
+export const VERSION = '0.7.0';
 export const CONFIG_SECTION = 'vlocity';
 export const OUTPUT_CHANNEL_NAME = 'Vlocity';
 export const NG_APP_NAME = 'Vlocode';