+++ src/models/webviewMessage.ts
@@ -1,4 +1,7 @@
-export default interface webviewMessage {
+/**
+ * Describes the message structure used for extening data between a webview and the extension.
+ */
+export default interface WebviewMessage {
     command: string;
-    data: any;
+    data?: any;
 }