+++ src/services/vlocodeService.ts
@@ -3,11 +3,12 @@ import * as vlocity from 'vlocity';
 import * as path from 'path';
 import * as process from 'process';
 import * as constants from '../constants';
-import VlocodeConfiguration from '../models/VlocodeConfiguration';
+import VlocodeConfiguration from 'models/VlocodeConfiguration';
 import VlocityDatapackService, * as vds from './vlocityDatapackService';
-import * as s from '../singleton';
-import * as l from '../loggers';
-import { ICommand } from '../commands/command';
+import * as serviceProvider from 'singleton';
+import { Logger } from 'loggers';
+import CommandRouter from './commandRouter';
+import { VlocodeCommand } from 'commands';
 
 export default class VlocodeService {  
 
@@ -16,14 +17,28 @@ export default class VlocodeService {
     private _context: vscode.ExtensionContext;
     private _datapackService: VlocityDatapackService;
     private _disposables: {dispose() : any}[] = [];
+    private _statusBar: vscode.StatusBarItem;
 
     constructor(context?: vscode.ExtensionContext, config?: VlocodeConfiguration) {
         if (!config) {
-            throw 'Cannot create VlocodeService without specifying a configuration object.';
+            throw new Error('Cannot create VlocodeService without specifying a configuration object.');
         }
         this._context = context;
         this.setConfig(config);
         context.subscriptions.push(this);
+    }
+
+    public showStatus(text: string, command: VlocodeCommand | string = undefined) : void {
+        if (!this._statusBar) {
+            this._statusBar = this.registerDisposable(vscode.window.createStatusBarItem(vscode.StatusBarAlignment.Left, 10));
+        }
+        this._statusBar.command = command;
+        this._statusBar.text = text;
+        this._statusBar.show();
+    }
+
+    public hideStatus() : void {
+        this._statusBar.hide();
     }
 
     public dispose() {
@@ -39,14 +54,29 @@ export default class VlocodeService {
     
     public setConfig(config: VlocodeConfiguration){
         this._config = config;
+        this.updateStatusBar(config);
+        this.registerDisposable(VlocodeConfiguration.watch(config, (c) => {
+            this._datapackService = null;
+            this.updateStatusBar(c);
+        }));
+    }
+
+    private updateStatusBar(config: VlocodeConfiguration) {
+        if (!config.sfdxUsername && !config.username) {
+            return this.showStatus(`$(gear) Vlocode: select org`, VlocodeCommand.selectOrg);
+        }
+        if (!config.projectPath) {
+            return this.showStatus(`$(file-directory) Vlocode: select project path`);
+        }
+        return this.showStatus(`$(cloud-upload) Vlocode ${config.sfdxUsername || config.username}`, VlocodeCommand.selectOrg);
     }
     
     public focusLog() : any {
         this.outputChannel.show(false);
     }
 
-    protected get logger() : l.Logger {
-        return s.get(l.Logger);
+    protected get logger() : Logger {
+        return serviceProvider.get(Logger);
     }
 
     public registerDisposable<T extends  {dispose() : any}>(disposable: T) : T {
@@ -54,36 +84,13 @@ export default class VlocodeService {
         return disposable;
     }
 
-    public registerCommand(...cmds: ICommand[]) : void {
-        cmds.forEach(cmd => {
-            this.logger.verbose(`Register commands ${cmd.name}`);
-            this.registerDisposable(vscode.commands.registerCommand(cmd.name, async (...args) => {     
-                try {
-                    s.get(VlocodeService).validateConfig();
-                    s.get(VlocodeService).validateSalesforce();
-                } catch (err) {
-                    this.logger.error(`${cmd.name}: ${err}`);
-                    return vscode.window.showErrorMessage(err, { modal: false }, { title: 'Open settings' }).then(r => 
-                        r === undefined || vscode.commands.executeCommand('workbench.action.openWorkspaceSettings', 'test'));
-                }
-                this.logger.verbose(`Invoke command ${cmd.name}`);
-                try {
-                    await cmd.execute.apply(cmd, args);
-                    this.logger.verbose(`Execution of command ${cmd.name} done`);
-                } catch(err) {
-                    this.logger.error(`Command execution resulted in error: ${err}`);
-                }
-            }))
-        });
-    }
-
-    public validateSalesforce() : void {
+    public validateSalesforce() : string {
         if (!this.datapackService.isVlocityPackageInstalled()) {
-            throw 'The Vlocity managed package is not installed on your Salesforce organization; select a different Salesforce organization or install the Vlocity managed package.';
+            return 'The Vlocity managed package is not installed on your Salesforce organization; select a different Salesforce organization or install the Vlocity managed package.';
         }
     }
 
-    public validateConfig() : void {
+    public validateConfig() : string {
         if (this.config.sfdxUsername) {
             // check for username and password
             if (this.config.username || this.config.password) {
@@ -91,17 +98,17 @@ export default class VlocodeService {
             }
         } else {
             if (!!this.config.username && !!this.config.password) {
-                throw 'Invalid condiguration - Please configure either a SFDX username or Salesforce login credential in order to use Vlocode.';
+                return 'Invalid condiguration - Please configure either a SFDX username or Salesforce login credential in order to use Vlocode.';
             } else if (this.config.instanceUrl) {
-                throw 'Invalid condiguration - Please configure the instance url for Salesforce; or use an SFDX authorized username.';
+                return 'Invalid condiguration - Please configure the instance url for Salesforce; or use an SFDX authorized username.';
             }
         }
         if (!this.config.projectPath) {
-            throw 'Invalid condiguration - Please configure the projectPath to point to the folder containing Vlocity datapack sources.';
+            return 'Invalid condiguration - Please configure the projectPath to point to the folder containing Vlocity datapack sources.';
         }
     }
 
-    private getVlocityJobOptions() : vlocity.JobOptions {
+    private get vlocityJobOptions() : vlocity.JobOptions {
         // for now this is a simple cast but in the future this migth change
         return <vlocity.JobOptions> (<any> this._config);
     }
@@ -111,7 +118,7 @@ export default class VlocodeService {
     }
 
     get datapackService(): VlocityDatapackService {
-        return this._datapackService || (this._datapackService = new VlocityDatapackService(this.getVlocityJobOptions()));
+        return this._datapackService || (this._datapackService = new VlocityDatapackService(this.vlocityJobOptions));
     }
 
     get outputChannel(): vscode.OutputChannel {